// src/interfaces/iDependency.ts
/**
 *
 * @export
 * @interface IDependency
 */
export interface IDependency {
  /**
   *
   * @type {string}
   * @memberof IDependency
   */
  'name': string;
  /**
   *
   * @type {string}
   * @memberof IDependency
   */
  'url': string;
  /**
   *
   * @type {string}
   * @memberof IDependency
   */
  'type': string;
  /**
   *
   * @type {string}
   * @memberof IDependency
   */
  'sha512sum': string;
}